+++ configs/rag.yaml
@@ -9,7 +9,7 @@ data:
 model:
   model_type: causal
   reranker_model: dlicari/distil-ita-legal-bert #"dlicari/Italian-Legal-BERT"
-  model_path: meta-llama/Meta-Llama-3-8B-Instruct #"dlicari/Italian-Legal-BERT-SC" #facebook/mbart-large-50 #Helsinki-NLP/opus-mt-it-en
+  model_path: "distilbert/distilgpt2" #"openai-community/gpt2-large" #meta-llama/Meta-Llama-3-8B-Instruct #"dlicari/Italian-Legal-BERT-SC" #facebook/mbart-large-50 #Helsinki-NLP/opus-mt-it-en
   adapter_path: #models/fine_tuned_models/opus-mt-it-en-v1/model
   tokenizer_path: #"dlicari/Italian-Legal-BERT-SC" #"meta-llama/Meta-Llama-3-8B-Instruct" #facebook/mbart-large-50   #models/fine_tuned_models/opus-mt-it-en-v1/tokenizer
   embedding_model: dlicari/distil-ita-legal-bert #"dlicari/Italian-Legal-BERT-SC" #"dlicari/Italian-Legal-BERT"  #intfloat/multilingual-e5-large-instruct
